refactor(tangle-dapp): Proper directory placement of `ApiDevStatsContainer`
