+++ apps/tangle-dapp/containers/ApiDevStatsContainer/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ApiDevStatsContainer';
