+++ apps/tangle-dapp/containers/Layout/Layout.tsx
@@ -3,7 +3,7 @@ import { getSideBarStateFromCookie } from '@webb-tools/webb-ui-components/next-u
 import React, { type FC, type PropsWithChildren } from 'react';
 
 import { Breadcrumbs, SideBar, SideBarMenu } from '../../components';
-import ApiDevStats from '../ApiDevStats';
+import ApiDevStatsContainer from '../ApiDevStatsContainer';
 import WalletAndChainContainer from '../WalletAndChainContainer/WalletAndChainContainer';
 import { WalletModalContainer } from '../WalletModalContainer';
 
@@ -37,7 +37,7 @@ const Layout: FC<PropsWithChildren> = ({ children }) => {
         <Footer isMinimal className="py-8" />
       </main>
 
-      {isDevelopment && <ApiDevStats />}
+      {isDevelopment && <ApiDevStatsContainer />}
     </div>
   );
 };
